refactor: extract OAuthTokenStorage from OAuthManager
Phase 4 file split. Pure extraction; bit-identical behaviour.

The Keychain + UserDefaults persistence concerns are now isolated in
`OAuthTokenStorage`. `OAuthManager` keeps the auth-flow orchestration but
delegates the actual `save/load/delete` to the storage type. `AccountTokens`
moves to the storage namespace (with a typealias in `OAuthManager` so the
auth-flow code still reads naturally).

Why
- `OAuthManager` is the most-touched service in this sprint (Phase 3c).
  Pulling the persistence layer out lets future auth-flow changes land
  without rebasing against storage diffs (and vice versa).
- Keychain layout decisions (only refresh_token persisted, access token
  re-derived on launch, expiry in UserDefaults) now have a dedicated home
  with a doc comment explaining the rationale.

Lines: OAuthManager 626 → ~603, new OAuthTokenStorage 66.

Tests: all 16 OAuth tests pass unchanged.
